+++ KalikoCMS.Engine/Data/Entities/SitePropertyDefinitionEntity.cs
@@ -5,14 +5,14 @@
  * Copyright (c) Fredrik Schultz
  * 
  * This library is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
+ * modify it under the terms of the GNU Lesser General public
  * License as published by the Free Software Foundation; either
  * version 3.0 of the License, or (at your option) any later version.
  * 
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
+ * Lesser General public License for more details.
  * http://www.gnu.org/licenses/lgpl-3.0.html
  */
 #endregion
@@ -20,20 +20,21 @@
 namespace KalikoCMS.Data.Entities {
     using System;
     using System.Collections.Generic;
-    using Core.Collections;
 
-    public class SiteEntity {
-        public virtual Guid SiteId { get; set; }
+    public class SitePropertyDefinitionEntity {
+        public virtual int PropertyId { get; set; }
+        public virtual Guid PropertyTypeId { get; set; }
         public virtual string Name { get; set; }
-        public virtual string Author { get; set; }
-        public virtual SortDirection ChildSortDirection { get; set; }
-        public virtual SortOrder ChildSortOrder { get; set; }
-        public virtual DateTime UpdateDate { get; set; }
-        
+        public virtual string Header { get; set; }
+        public virtual bool ShowInAdmin { get; set; }
+        public virtual int SortOrder { get; set; }
+        public virtual string Parameters { get; set; }
+        public virtual bool Required { get; set; }
+        public virtual PropertyTypeEntity PropertyType { get; set; }
         public virtual IList<SitePropertyEntity> SiteProperties { get; private set; }
 
-        public SiteEntity() {
-            SiteProperties =new List<SitePropertyEntity>();
+        public SitePropertyDefinitionEntity() {
+            SiteProperties = new List<SitePropertyEntity>();
         }
     }
 }